Agentic_RAG/requirements.txt
openai>=1.30.0
faiss-cpu>=1.8.0
boto3>=1.34.0
numpy>=1.26.0
python-dotenv>=1.0.0
